requirements/requirements.txt
boto3==1.4.7
botocore==1.7.33
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.1.2
Django==1.10.8
django-allauth==0.33.0
docutils==0.14
enum34==1.1.6
functools32==3.2.3.post2
futures==3.1.1
idna==2.6
ipython-genutils==0.2.0
jmespath==0.9.3
jsonschema==2.6.0
jupyter-core==4.3.0
nbformat==4.4.0
oauthlib==2.0.4
plotly==2.1.0
python-dateutil==2.6.1
python-openid==2.2.5
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
s3transfer==0.1.11
six==1.11.0
traitlets==4.3.2
urllib3==1.22